Main8.java
This Java program demonstrates the use of instance and static variables and methods. The Sample class has an instance variable x and a static variable y. The display() method shows how instance and static variables can be accessed through an object. The static method display1() accesses the instance variable by creating a new object and directly accesses the static variable. In the main method, both instance and static methods are called to illustrate their behavior.