Main6.java
This Java program demonstrates the use of a static method in an interface and implementation of an abstract method. The interface one contains a static method display() and an abstract method show(). The class sample implements one and provides the body for show(), but also defines a method named sample(), which is not a constructor due to the void return type.